增加收入支出功能;修改mapper文件if标签问题

+++ gym_java/src/main/java/edu/gdou/gym_java/entity/model/EquipmentRentStandard.java
@@ -9,6 +9,9 @@ import io.swagger.annotations.ApiModelProperty;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+
+import java.util.Date;
+
 /**
  * <p>
  * 器材费用标准表
@@ -37,7 +40,7 @@ public class EquipmentRentStandard {
 
     @ApiModelProperty("器材租用费用")
     @TableField(value="eid")
-    private Integer price;
+    private Double price;
 
     @ApiModelProperty("器材租用时间")
     @TableField(value="rentTime")
